QuestManager reorg - Extract AssignedQuest
